refactor(test): template metadata tests should test new class, not private functions
